Improvements related to android logging
There is no need for the extra defines. Now it's managed in runtime with the new function log_android_setup.
Also added log_add_filePath which similar to log_add_fp adds a file to log to, but instead, it is opened and closed on each log call. While it's not performant, it adds some assurance that the fille will be closed if the application crashes for some reason. I needed this for some android debugging.