Callbacks can now store a string for future use
When udata == NULL then save_str will be used in place of udata on callbacks.
In some cases it was convenient to have a FILE descriptor or something else on user data, but having to allocate a char* in memory application side just so that the logging librery can access it all the time is bothersome. Instead I think it's better to have the option to save a small string (256 characters) on the callback struct, and if udata is a NULL pointer, assume that said string should be used as the udata. This is being used so far for storing the application identifier string when logging in android (for logcat) as well as saving a path to a file to append to on every log call, also usefull in android when I never know when a crash will happen and won't be able to properly close the FILE descriptor, so instead I close it all the time.